+++ .github/scripts/run_code_in_mdfile.py
@@ -44,7 +44,7 @@ def process_file(file, language, version):
             rust_script_contents = f"""//! ```cargo
         //! [dependencies]
         //! lace = {{ path = ".", version="{version}" }}
-        //! polars = {{ version = "=0.32.0", features=["csv"] }}
+        //! polars = {{ version = "0.33", features=["csv"] }}
         //! rand = "0.8"
         //! rand_xoshiro = "0.6"
         //! ```